[DIP] Invertendo as dependências da classe Cotuba
