+++ cotuba/src/main/java/cotuba/application/Cotuba.java
@@ -3,8 +3,11 @@ package cotuba.application;
 import cotuba.domain.Capitulo;
 import cotuba.domain.Ebook;
 import cotuba.epub.GeradorEPUB;
+import cotuba.epub.GeradorEPUBImpl;
 import cotuba.md.RenderizadorMDParaHTML;
+import cotuba.md.RenderizadorMDParaHTMLImpl;
 import cotuba.pdf.GeradorPDF;
+import cotuba.pdf.GeradorPDFImpl;
 
 import java.nio.file.Path;
 import java.util.List;
@@ -13,7 +16,7 @@ public class Cotuba {
 
   public void executa(String formato, Path diretorioDosMD, Path arquivoDeSaida) {
 
-    RenderizadorMDParaHTML renderizador = new RenderizadorMDParaHTML();
+    RenderizadorMDParaHTML renderizador = new RenderizadorMDParaHTMLImpl();
     List<Capitulo> capitulos = renderizador.renderiza(diretorioDosMD);
 
     Ebook ebook = new Ebook();
@@ -23,12 +26,12 @@ public class Cotuba {
 
     if ("pdf".equals(formato)) {
 
-      GeradorPDF geradorPDF = new GeradorPDF();
+      GeradorPDF geradorPDF = new GeradorPDFImpl();
       geradorPDF.gera(ebook);
 
     } else if ("epub".equals(formato)) {
 
-      GeradorEPUB geradorEPUB = new GeradorEPUB();
+      GeradorEPUB geradorEPUB = new GeradorEPUBImpl();
       geradorEPUB.gera(ebook);
 
     } else {
